Extract Steps components from batch inventory
